added create fav test after relaunch
