+++ app/src/androidTest/java/com/freshdigitable/udonroad/MockTwitterApiComponent.java
@@ -14,7 +14,5 @@ import dagger.Component;
 @Singleton
 @Component(modules = MockTwitterApiModule.class)
 public interface MockTwitterApiComponent extends TwitterApiComponent {
-  void inject(MainActivityInstTest mainActivityInstTest);
-
-  void inject(MainActivityResumeInstTest mainActivityResumeInstTest);
+  void inject(MainActivityInstTestBase mainActivityResumeInstTest);
 }
